Rename generic utility files to domain-specific names
- Rename helpers.ts → cached-size-analysis.ts (loop-size-hoisting)
- Rename utils.ts → diagnostic-fix-builder.ts (feather)
- Rename common.ts → conflict-resolution.ts (identifier-case)
- Update all import references
- Apply formatter to ensure consistency
