+++ src/cli/src/cli-core/commander-types.ts
@@ -29,8 +29,7 @@ export interface CommanderAddCommandOptions {
 export type CommanderActionHandler = (...args: Array<unknown>) => unknown;
 
 export interface CommanderCommandHost
-    extends CommanderUsageProvider,
-        CommanderLifecycle {
+    extends CommanderUsageProvider, CommanderLifecycle {
     addCommand?: (
         command: CommanderCommandHost,
         options?: CommanderAddCommandOptions
