+++ src/cli/src/cli-core/format-command-options.ts
@@ -28,8 +28,7 @@ export interface CollectFormatCommandOptionsParameters {
 }
 
 export interface FormatCommandOptionsResult
-    extends FormatCommandSampleLimits,
-        ResolvedPrettierConfiguration {
+    extends FormatCommandSampleLimits, ResolvedPrettierConfiguration {
     targetPathInput: unknown;
     targetPathProvided: boolean;
     extensions: Array<string>;
