+++ src/core/src/comments/doc-comment/service/synthetic-generation.ts
@@ -994,8 +994,8 @@ function applyImplicitNameOverride({
     const resolvedFallbackCanonical = fallbackCanonical;
     const shouldOverrideImplicitName = Boolean(
         canonicalOrdinal &&
-            canonicalOrdinal !== resolvedFallbackCanonical &&
-            canonicalOrdinal !== canonicalImplicit
+        canonicalOrdinal !== resolvedFallbackCanonical &&
+        canonicalOrdinal !== canonicalImplicit
     );
 
     if (!shouldOverrideImplicitName) {
