+++ src/plugin/src/transforms/loop-size-hoisting/cached-size-analysis.ts
