+++ src/plugin/src/transforms/loop-size-hoisting/index.ts
@@ -1,2 +1,2 @@
-export * from "./helpers.js";
+export * from "./cached-size-analysis.js";
 export { hoistLoopLengthBounds } from "./hoist-loop-length.js";
